+++ lastversion/__about__.py
@@ -1,3 +1,3 @@
 """Package metadata"""
-__version__ = '3.2.1'
+__version__ = '3.3.0'
 __self__ = "dvershinin/lastversion"